break the stereo eval function to better organize it
